Enhance code quality and error handling across multiple files. Updated .clang-tidy configuration to include additional checks for member variable visibility and cognitive complexity. Changed ErrorCode enum to use uint8_t for better memory efficiency. Improved exception handling in main CLI entry point to provide clearer error messages. Refined SHA256 implementation for consistency in byte handling during hashing operations.

+++ include/blockchain/error.hpp
@@ -12,7 +12,7 @@ namespace blockchain {
 // Stable, inspectable error categories shared across the codebase. These are
 // intended for tests to assert on, so values must remain meaningful and the
 // set must only grow in an additive way.
-enum class ErrorCode : std::uint16_t {
+enum class ErrorCode : std::uint8_t {
   kParseError = 1,
   kInvalidMessage,
   kInvalidTransaction,
